+++ lib/sound_util.rb
@@ -6,4 +6,7 @@ module SoundUtil
   class Error < StandardError; end
 
   autoload :CLI, "sound_util/cli"
+  autoload :Filter, "sound_util/filter"
+  autoload :Generator, "sound_util/generator"
+  autoload :Wave, "sound_util/wave"
 end