+++ lib/sound_util/generator.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module SoundUtil
+  module Generator
+    autoload :Tone, "sound_util/generator/tone"
+  end
+end